requirements.txt
inquirer
pydobot
pyserial
typer
